requirements.txt
chromadb==1.3.5
google_search_results==2.4.2
jsonlines==4.0.0
networkx==3.4.2
numpy==2.4.1
openai==2.8.1
pandas==2.3.3
Pillow==12.0.0
Requests==2.32.5
rouge_score==0.1.2
sentence_transformers==3.4.1
serpapi==0.1.5
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
-e git+https://github.com/lotus-data/lotus.git@59acc83886a5901a76025703538259a40cb945b3#egg=lotus_ai